services/template-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.11.7

# Database dependencies - SQLAlchemy 2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.7
alembic==1.16.2

# HTTP client for external services
httpx==0.28.1

# Utilities
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1

# Development and testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==4.1.0
black==23.9.1
isort==5.12.0
flake8==6.0.0